Nice little Rotary Knob processor
Faderfox UC4 has rotary knobs that function as buttons - so with great assistance from the Pure Data community, especially Kai PreuB, got it working. This patch switches output from being a rotary knob controlling 0.1 fidelity to 0.001 fidelity - all you have to do is set up the first inlet as the knob and the 2nd inlet as the button - so feed in (ctlin x) and (ctlin y) to it.